Redirect to our own view after payments
We're not redirecting to a view of our own after a user completes a
payment. This lets us capture the data that MP sends via GET (but does
not send as a proper notification) immediately.

This view will then forward the payee to a configurable view with the
Notification id.